requirements.txt
fastapi 
uvicorn 
psycopg2-binary 
coremltools 
sentence-transformers 
pgvector
requests
python-dotenv
flask